feat [Backend]: authorize membership endpoints

+++ api/app/Http/Controllers/MembershipController.php
@@ -11,6 +11,12 @@ class MembershipController extends Controller
 {
     public function index(Request $request)
     {
+        $membership = Membership::fromRequest($request)
+            ->where('user_id', auth()->id())
+            ->first();
+            
+        $this->authorize('viewAny', [Membership::class, $membership->membershipable]);
+
         $role = $request->get('role');
         $search = $request->get('search');
 
@@ -26,6 +32,8 @@ class MembershipController extends Controller
 
     public function update(UpdateMembershipRequest $request, string $type, int $id, Membership $membership)
     {
+        $this->authorize('update', $membership);
+
         $membership->syncRoles([$request->validated('role')]);
 
         return response()->noContent();
@@ -33,6 +41,8 @@ class MembershipController extends Controller
 
     public function delete(string $type, int $id, Membership $membership)
     {
+        $this->authorize('delete', $membership);
+
         $membership->delete();
 
         return response()->noContent();
